+++ src/settings/_index.njk
@@ -19,26 +19,6 @@
   <h2>
     Manage Data
   </h2>
-  <p>
-    If you use this site for free, your data is completely private. However, that
-    means it isn't backed up in the cloud. If you want to access your data from
-    multiple devices or keep backups, you'll need to export or import data from
-    here.
-  </p>
-
-  <h3>
-    Export Data
-  </h3>
-  <p>
-    In case you drop your phone in the toilet, it's a good idea to back up your
-    data regularly. You can export a file with all your data from here, then use
-    whatever other system you like for saving backups.
-  </p>
-  <div>
-    <button type="button" data-export-data>
-      Export Data
-    </button>
-  </div>
 
   <h3>
     Import Data
