+++ src/settings/export-data.njk
@@ -15,8 +15,19 @@
       Go Back
     </a>
   </div>
+
+  <p>
+    In case you drop your phone in the toilet, it's a good idea to back up your
+    data regularly. You can export a file with all your data from here, then use
+    whatever other system you like for saving backups.
+  </p>
+  <div>
+    <button type="button" data-export-data>
+      Export Data
+    </button>
+  </div>
 {% endblock %}
 
 {% block scripts %}
-  <script src="/scripts/settings/export-data.js"></script>
+  <script src="/scripts/settings/export-data.mjs" type="module"></script>
 {% endblock %}
